Moved test classes to their respective packages
