+++ src/test/java/com/gloot/springbootcodetest/leaderboard/repository/LeaderboardRepositoryTest.java
@@ -1,4 +1,4 @@
-package com.gloot.springbootcodetest.leaderboard;
+package com.gloot.springbootcodetest.leaderboard.repository;
 
 import com.gloot.springbootcodetest.SpringBootComponentTest;
 import com.gloot.springbootcodetest.leaderboard.entity.LeaderboardEntryEntity;
