+++ src/test/java/com/gloot/springbootcodetest/leaderboard/service/LeaderboardServiceTest.java
@@ -1,4 +1,4 @@
-package com.gloot.springbootcodetest.leaderboard;
+package com.gloot.springbootcodetest.leaderboard.service;
 
 import com.gloot.springbootcodetest.SpringBootComponentTest;
 import com.gloot.springbootcodetest.leaderboard.dto.LeaderboardEntryDto;
